feat(contracts): stawka kandydata → Numeric(10,2) (import zamówień Erste)
Migracje 0147 (VeloBank #593, Polkomtel #594) przeszły na Numeric(10,2) stawkę
KLIENTA, ale stawka KANDYDATA została Integer. Import zamówień Erste ma kandydata
ze stawką godzinową z połówką (157.5 PLN/h) — nie da się jej zapisać jako Integer.

Migracja 0148:
- scala dwa równoległe heady 0147 w jeden head (down_revision = krotka),
- przełącza rodzinę stawki kandydata na Numeric(10,2) symetrycznie do strony
  klienta: contracts.rate_candidate, contract_candidate_rates.rate (harmonogram),
  candidate_rate_history.rate (historia).

Modele + schematy stawki kandydata → Decimal/float (konwencja per-plik jak na
main: Decimal w client_order/new_contractor_order, float w contract.py).
contract_amendment.new_rate_* też dziesiętne. Logika marży zwraca Decimal→Numeric.

Konwersja int→numeric bezstratna, in-place. Frontend bez zmian (formatCurrency/
fmtMoney renderują połówki; ścieżka Decimal istnieje od strony klienta).

+++ backend/app/models/rate_history.py
@@ -7,9 +7,10 @@ which client on what contract type" before making a new offer.
 
 import enum
 from datetime import date
+from decimal import Decimal
 from typing import Optional
 
-from sqlalchemy import Date, Enum, ForeignKey, Integer, String, Text
+from sqlalchemy import Date, Enum, ForeignKey, Integer, Numeric, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.database import Base
@@ -39,7 +40,7 @@ class RateHistory(Base, TimestampMixin):
         ForeignKey("jobs.id"), nullable=True, index=True
     )
 
-    rate: Mapped[int] = mapped_column(Integer, nullable=False)
+    rate: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     currency: Mapped[str] = mapped_column(String(3), default="PLN", nullable=False)
     contract_type: Mapped[ContractType] = mapped_column(
         Enum(ContractType, name="contracttype"), nullable=False
